chore: Rename folder to `multiSelect`
