+++ mobile/src/navigation/routes.tsx
@@ -58,7 +58,7 @@ import { useTabsByVisibility } from "~/stores/Preference/hooks";
 
 import { capitalize } from "~/utils/string";
 import type { Tab } from "~/stores/Preference/types";
-import { TrackMultiSelectListeners } from "~/modules/media/track/components/TrackMultiSelectListeners";
+import { TrackMultiSelectListeners } from "~/modules/media/multiSelect/components/TrackMultiSelectListeners";
 import { BottomActions } from "./components/BottomActions";
 import { DeferredRender } from "./components/DeferredRender";
 import { TopAppBar } from "./components/TopAppBar";
